Refactor ProposerDuty
